Update negotiator
preferredMediaType -> mediaType

The behavior for requests with no accept header is to default to a
header value of "*/*" (see [1]). This means that the first type handler
will be used instead of the default handler. The default handler will
only be used if the accept header does not match any given types.

[1] https://github.com/jshttp/negotiator/commit/b49537f407d2c1ae7a4c02963d1aec8af78c8cc0